% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumr.R
\name{cusumr}
\alias{cusumr}
\title{Perform cumulative sum analysis on binary dataset}
\usage{
cusumr(
  events_outcomes,
  acceptable_rate = 0.2,
  unacceptable_rate = 0.4,
  type1_error_rate = 0.1,
  type2_error_rate = 0.1,
  learning = TRUE,
  reset = TRUE
)
}
\arguments{
\item{events_outcomes}{vector = [n_samples of 0s and 1s]. Binary Integer values. If intergers are not 0 or 1, then the FALSE alarms should be explicitly given.}

\item{acceptable_rate}{numeric, default=0.2. Number between 0 and 1. Acceptable success rate of the process being monitored.}

\item{type1_error_rate}{numeric, default=0.1. Number between 0 and 1. A false positive or type 1 error rate}

\item{type2_error_rate}{numeric, default=0.2 Number between 0 and 1. A false negative or type 2 error rate}

\item{learning}{boolean, optional (default=True). Whether to start from learning phase or from monitoring phase.}

\item{reset}{boolean, optional (default=True). Whether to reset score when the score hits the decision limit in monitoring phase. If yes, the cusum score will start at zero again and restart monitoring.}
}
\description{
Perform cumulative sum analysis on binary dataset
}
